refactor : trainer profile 분리
