+++ src/main/java/com/example/healthgenie/boundedContext/trainer/profile/dto/ProfileResponseDto.java
@@ -1,7 +1,7 @@
-package com.example.healthgenie.boundedContext.trainer.dto;
+package com.example.healthgenie.boundedContext.trainer.profile.dto;
 
-import com.example.healthgenie.boundedContext.trainer.entity.TrainerInfo;
-import com.example.healthgenie.boundedContext.trainer.entity.TrainerPhoto;
+import com.example.healthgenie.boundedContext.trainer.photo.entity.TrainerPhoto;
+import com.example.healthgenie.boundedContext.trainer.profile.entity.TrainerInfo;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
