+++ src/main/java/com/example/healthgenie/boundedContext/trainer/profile/dto/ProfileSliceResponse.java
@@ -1,6 +1,6 @@
-package com.example.healthgenie.boundedContext.trainer.dto;
+package com.example.healthgenie.boundedContext.trainer.profile.dto;
 
-import com.example.healthgenie.boundedContext.trainer.entity.TrainerInfo;
+import com.example.healthgenie.boundedContext.trainer.profile.entity.TrainerInfo;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
