+++ src/main/java/com/example/healthgenie/boundedContext/trainer/profile/entity/TrainerInfo.java
@@ -1,6 +1,7 @@
-package com.example.healthgenie.boundedContext.trainer.entity;
+package com.example.healthgenie.boundedContext.trainer.profile.entity;
 
 import com.example.healthgenie.base.entity.BaseEntity;
+import com.example.healthgenie.boundedContext.trainer.photo.entity.TrainerPhoto;
 import com.example.healthgenie.boundedContext.user.entity.User;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
