+++ src/test/java/com/example/healthgenie/util/TestSyUtils.java
@@ -20,11 +20,11 @@ import com.example.healthgenie.boundedContext.todo.dto.TodoRequestDto;
 import com.example.healthgenie.boundedContext.todo.dto.TodoUpdateRequest;
 import com.example.healthgenie.boundedContext.todo.entity.Todo;
 import com.example.healthgenie.boundedContext.todo.repository.TodoRepository;
-import com.example.healthgenie.boundedContext.trainer.dto.ProfileRequestDto;
-import com.example.healthgenie.boundedContext.trainer.entity.TrainerInfo;
-import com.example.healthgenie.boundedContext.trainer.entity.TrainerPhoto;
-import com.example.healthgenie.boundedContext.trainer.repository.TrainerProfilePhotoRepository;
-import com.example.healthgenie.boundedContext.trainer.repository.TrainerProfileRepository;
+import com.example.healthgenie.boundedContext.trainer.photo.entity.TrainerPhoto;
+import com.example.healthgenie.boundedContext.trainer.photo.repository.TrainerProfilePhotoRepository;
+import com.example.healthgenie.boundedContext.trainer.profile.dto.ProfileRequestDto;
+import com.example.healthgenie.boundedContext.trainer.profile.entity.TrainerInfo;
+import com.example.healthgenie.boundedContext.trainer.profile.repository.TrainerProfileRepository;
 import com.example.healthgenie.boundedContext.user.entity.User;
 import com.example.healthgenie.boundedContext.user.entity.enums.Role;
 import com.example.healthgenie.boundedContext.user.repository.UserRepository;
